Promote the shared no-other-transactions scan into the warrant checks file
